Module Reader.Make

The implementation of read effects.

Signatures of read effects.

Parameters

module P : Param

Signature

include Param with type env = P.env

type env = P.env

The type of environments.

val read : unit -> env

Read the environment.

val scope : ( env -> env ) -> ( unit -> 'a ) -> 'a

scope f t runs the thunk t under the new environment that is the result of applying f to the current environment.

val run : env:env -> ( unit -> 'a ) -> 'a

run t runs the thunk t which may perform reading effects.