Module type Sequencer.S

Signatures of sequencing effects.

type elt

The type of elements.

val yield : elt -> unit

Yield the element.

val run : (unit -> unit) -> elt Stdlib.Seq.t

run t runs the thunk t which may perform sequencing effects.

val register_printer : ([ `Yield of elt ] -> string option) -> unit

register_printer p registers a printer p via Printexc.register_printer to convert unhandled internal effects into strings for the OCaml runtime system to display. Ideally, all internal effects should have been handled by run and there is no need to use this function, but when it is not the case, this function can be helpful for debugging. The functor Sequencer.Make always registers a simple printer to suggest using run, but you can register new ones to override it. The return type of the printer p should return Some s where s is the resulting string, or None if it chooses not to convert a particular effect. The registered printers are tried in reverse order until one of them returns Some s for some s; that is, the last registered printer is tried first. Note that this function is a wrapper of Printexc.register_printer and all the registered printers (via this function or Printexc.register_printer) are put into the same list.

The input type of the printer p is a variant representation of the internal effects used in this module. They correspond to the effects trigger by yield. More precisely, `Yield elt corresponds to the effect triggered by yield elt.