Module Unmonad.Make

The implementation of monad effects.

Signatures of monad effects.

Parameters

module M : Monad

Signature

type 'a t = 'a M.t

The monad.

val perform : 'a t -> 'a

Perform an monadic operation.

val run : ( unit -> 'a ) -> 'a t

run t runs the thunk t which may perform monad effects, and then returns the corresponding monadic expression.